# swagger.yaml
openapi: "3.0.0"
info:
  title: learn express and swagger
  description: A course segment about writing documentation
  version: 1.1.0
  contact: 
    email: ananjaygurjar92@gmail.com
    url: "https://www.google.com" #we must use ""(double quotes) in case of the url

# servers:
#   - url: "https://localhost:4000/api/v1"
#     description: for local host - secure
#   - url: "http://localhost:4000/api/v1"
#     description: for local host - regular

#using variables to describe the servers

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: for local host 
    variables:
      version:
        enum:
          - v1
          - v2
        default: v1
      protocol:
        enum:
          - http
          - https
        default: http

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

#creating a route which will return a string

paths:
  /lco:
    get:
      tags:
        - String
      summary: returns a greet message from lco
      responses:
        200:
          description: All good and success
          content: 
            application/json:
              schema:
                type: string
                example: "myString"
        400:
          description: Bad Request
        500:
          description: Internal server error
  /lcoobject:
    get:
      tags:
        - Object
      summary: returns a unique course
      responses:
        200:
          description: All good and success
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  id: 
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad Request
        500:
          description: Internal server error

  /courses:
    get:
      tags:
        - Array
      summary: returns all courses
      responses:
        200:
          description: All good and success
          content: 
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties: 
                    id: 
                      type: string
                    name:
                      type: string
                    price:
                      type: number
        400:
          description: Bad Request
        500:
          description: Internal server error

  /mycourse/{courseId}:
    get:
      tags:
        - String
      summary: adds a new course to existing courses
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: courseId
          in: path
          required: true
          default: 22
          schema: 
            type: string
      responses:
        200:
          description: All good and success
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  id: 
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        400:
          description: Bad Request
        500:
          description: Internal server error

  /addCourse:
      post:
        tags:
          - Array
        summary: returns course based on request id
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties: 
                  id: 
                    type: string
                  name:
                    type: string
                  price:
                    type: number
        responses:
          200:
            description: All good and success
            content: 
              application/json:
                schema:
                  type: boolean
          400:
            description: Bad Request
          500:
            description: Internal server error

#computed url - http://localhost:4000/api/v1/courseQuery?location=delhi&device=web

  /courseQuery:
      get:
        tags:
          - String
        summary: trying to learn about query
        parameters:
          - name: location
            in: query
            required: true
            schema:
              type: string
              enum: [delhi, london, jaipur]

          - name: device
            in: query
            required: true
            schema:
              type: string
              enum: [web, mobile]
        responses:
          200:
            description: All good and success
            content: 
              application/json:
                schema:
                  type: object
                  properties:
                    location:
                      type: string
                    device:
                      type: string
          400:
            description: Bad Request
          500:
            description: Internal server error

  /courseUpload:
      post:
        tags:
          - String
        summary: Uploading course image
        requestBody:
          content:
            multipart/form-data:
              schema:
                type: object
                properties:
                  file: 
                    type: string
                    format: binary
        responses:
          200:
            description: All good and success
          400:
            description: Bad Request
          500:
            description: Internal server error